Rename package using _ for C++/Python interop
