+++ testbed-ws/src/diff_drive_poc/package.xml
@@ -1,13 +1,14 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>diff-drive-poc</name>
+  <name>diff_drive_poc</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="yashrahmed@gmail.com">yashrahmed</maintainer>
   <license>TODO: License declaration</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_cmake_python</buildtool_depend>
 
   <depend>hardware_interface</depend>
   <depend>pluginlib</depend>
