+++ testbed-ws/src/diff_drive_poc/resources/gazebo/diffbot.gazebo.xacro
