+++ testbed-ws/src/diff_drive_poc/scripts/arduino-motor-driver.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+print("hello")
